\name{GeneF}


\alias{GeneF}


\docType{package}


\title{
Package for Generalized F-statistics
}


\description{
This package implements several generalized \eqn{F}{F}-statistics. The current version includes a generalized \eqn{F}{F}-statistic based on the flexible isotonic/monotonic regression or order restricted hypothesis testing.
}


\details{
\tabular{ll}{
Package: \tab GeneF\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-07-18\cr
License: \tab GPL version 2 or newer\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}


\author{
Yinglei Lai

Maintainer: ylai@gwu.edu
}


\keyword{htest}