\name{GeneF}


\alias{GeneF}


\docType{package}


\title{
Package for Generalized F-Statistics
}


\description{
Implementation of several generalized \eqn{F}{F}-statistics. The current version includes a generalized \eqn{F}{F}-statistic based on the flexible isotonic/monotonic regression or order restricted hypothesis testing. Based on: Y. Lai (2011) <doi:10.1371/journal.pone.0019754>.
}


\details{
\tabular{ll}{
Package: \tab GeneF\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2022-05-06\cr
License: \tab GPL version 2 or newer\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}


\author{
Yinglei Lai

Maintainer: ylai@gwu.edu
}


\keyword{htest}